+++ instapy/browser.py
@@ -7,7 +7,7 @@ from selenium.webdriver.common.proxy import Proxy, ProxyType
 from selenium.webdriver.common.keys import Keys
 from selenium.webdriver.firefox.options import Options as Firefox_Options
 from selenium.webdriver import Remote
-
+from fake_useragent import UserAgent, FakeUserAgentError
 
 # general libs
 import re
@@ -19,6 +19,7 @@ from .util import highlight_print
 from .util import emergency_exit
 from .util import get_current_url
 from .util import check_authorization
+from .util import web_address_navigator
 from .settings import Settings
 from .file_manager import get_chromedriver_location
 
@@ -40,9 +41,24 @@ def set_selenium_local_session(proxy_address,
     browser = None
     err_msg = ''
 
+    # define fallback useragent
+    user_agent = (
+        'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 '
+        '(KHTML, like Gecko) Chrome/74.0.3729.169 Safari/537.36'
+    )
+
+    # try to fetch latest user agent
+    try:
+        ua = UserAgent(fallback = user_agent)
+        user_agent = ua.firefox if use_firefox else ua.chrome
+    except FakeUserAgentError:
+        print('Latest user agent currently not reachable. Using fallback.')
+
+    # keep user_agent
+    Settings.user_agent = user_agent
     if use_firefox:
         firefox_options = Firefox_Options()
-        
+
         if headless_browser:
             firefox_options.add_argument('-headless')
 
@@ -54,6 +70,7 @@ def set_selenium_local_session(proxy_address,
 
         # set English language
         firefox_profile.set_preference('intl.accept_languages', 'en')
+        firefox_profile.set_preference('general.useragent.override', user_agent)
 
         if disable_image_load:
             # permissions.default.image = 2: Disable images load,
@@ -70,6 +87,9 @@ def set_selenium_local_session(proxy_address,
                                            proxy_address)
             firefox_profile.set_preference('network.proxy.ssl_port',
                                            proxy_port)
+
+        # mute audio while watching stories
+        firefox_profile.set_preference('media.volume_scale', '0.0')
 
         browser = webdriver.Firefox(firefox_profile=firefox_profile,
                                     options=firefox_options)
@@ -91,22 +111,18 @@ def set_selenium_local_session(proxy_address,
         chrome_options.add_argument('--dns-prefetch-disable')
         chrome_options.add_argument('--lang=en-US')
         chrome_options.add_argument('--disable-setuid-sandbox')
+        chrome_options.add_argument('--no-sandbox')
+        chrome_options.add_argument('--user-agent={user_agent}'
+                                    .format(user_agent = user_agent))
 
         # this option implements Chrome Headless, a new (late 2017)
         # GUI-less browser. chromedriver 2.9 and above required
         if headless_browser:
             chrome_options.add_argument('--headless')
-            chrome_options.add_argument('--no-sandbox')
 
             if disable_image_load:
                 chrome_options.add_argument(
                     '--blink-settings=imagesEnabled=false')
-
-            # replaces browser User Agent from "HeadlessChrome".
-            # iPhone + Safari
-            user_agent = "Mozilla/5.0 (iPhone; CPU iPhone OS 10_2 like Mac OS X) AppleWebKit/602.3.12 (KHTML, like Gecko) Version/10.0 Mobile/14C92 Safari/602.1"
-            chrome_options.add_argument('user-agent={user_agent}'
-                                        .format(user_agent=user_agent))
 
         capabilities = DesiredCapabilities.CHROME